+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Added a collection of theme colors to be available in the editor.
 - Added alternative style for Media & Text block.
+- Updated padding and font size for Button block.
+- Removed border radius option for Button block.
+- Changed default color for Button block.
 
 ## 0.1.0